removed GetActuatorStatus for now. changed to new kill topic
