+++ test/riptide_autonomy/bt_actions/TestGetBoolTopic.cpp
@@ -1,6 +1,6 @@
 #include "autonomy_test/autonomy_testing.hpp"
 
-TEST_F(BtTest, test_GetActuatorStatus) {
+TEST_F(BtTest, test_GetBoolTopic) {
     //TODO: define a test here and delete below line when you do
     GTEST_SKIP();
 }
