+++ trees/TestTree.xml
@@ -26,9 +26,8 @@
     </BehaviorTree>
     <BehaviorTree ID="BehaviorTree">
         <Sequence>
-            <Action ID="Info" message="hello!"/>
-            <Action ID="Wait" seconds=""/>
-            <Action ID="Info" message="goodbye!"/>
+            <SubTree ID="WaitForKillSwitch"/>
+            <Action ID="Info" message="LETS GOOOO"/>
         </Sequence>
     </BehaviorTree>
     <BehaviorTree ID="BuoyTree">
@@ -221,7 +220,7 @@
     </BehaviorTree>
     <BehaviorTree ID="LessCoolTorpedoAlign">
         <Sequence name="Less cool align">
-            <Action ID="TransformPose" from_frame="tempest/torpedo_link" name="get torpedo in origin frame" op="0" or="0" out_op="" out_or="" out_oy="" out_x="" out_y="{torp_y}" out_z="{torp_z}" oy="0" to_frame="tempest/base_link" x="0" y="0" z="0"/>
+            <Action ID="TransformPose" from_frame="talos/torpedo_link" name="get torpedo in origin frame" op="0" or="0" out_op="" out_or="" out_oy="" out_x="" out_y="{torp_y}" out_z="{torp_z}" oy="0" to_frame="talos/base_link" x="0" y="0" z="0"/>
             <Action ID="Math" a="0" b="{torp_z}" operator="-" out="{prop_z}"/>
             <Action ID="TransformPose" from_frame="torpedoHole_frame" op="0" or="0" out_op="{op}" out_or="{or}" out_oy="{oy}" out_x="{x}" out_y="{y}" out_z="{z}" oy="3.1415" to_frame="world" x="0.5" y="{torp_y}" z="{prop_z}"/>
             <SubTree ID="Move" __shared_blackboard="false" op="op" or="or" oy="oy" x="x" y="y" z="z"/>
@@ -267,7 +266,7 @@
                 <ForceSuccess>
                     <Sequence>
                         <Action ID="GetOdometry" op="{cr}" or="{cp}" oy="{cy}" x="" y="" z=""/>
-                        <Action ID="Error" message="Couldn't bring Tempest to the desired orientation! Wanted {r}, {p}, {y} but got {cr}, {cp}, {cy}. Moving on anyway."/>
+                        <Action ID="Error" message="Couldn't bring talos to the desired orientation! Wanted {r}, {p}, {y} but got {cr}, {cp}, {cy}. Moving on anyway."/>
                     </Sequence>
                 </ForceSuccess>
             </Fallback>
@@ -345,7 +344,7 @@
             <Action ID="Wait" seconds="5"/>
             <RetryUntilSuccessful num_attempts="-1">
                 <Sequence>
-                    <Action ID="GetSwitchState" aux_switch_inserted="" kill_switch_inserted="kill_switch_inserted"/>
+                    <Action ID="GetBoolTopic" topic="state/kill" value="{kill_switch_inserted}"/>
                     <Condition ID="IsTrue" value="{kill_switch_inserted}"/>
                     <Action ID="Info" message="KILL SWITCH INSERTED, TREE STARTING IN 5 SECONDS"/>
                 </Sequence>
@@ -415,11 +414,29 @@
             <input_port name="z1" required="false"/>
             <input_port name="z2" required="false"/>
         </Action>
+        <Action ID="CallSetBoolService">
+            <input_port name="data" required="true">boolean value to set the thing to</input_port>
+            <input_port name="srv_name" required="true">Name of the service to call</input_port>
+            <input_port name="time_limit_secs" required="true">time limit for service after the call is made. Another second may be consumed waiting for it to come up</input_port>
+        </Action>
+        <Action ID="CallTriggerService">
+            <input_port name="srv_name" required="true"/>
+            <input_port name="time_limit_secs" required="true"/>
+        </Action>
         <Condition ID="CompareNums">
             <input_port name="a" required="false"/>
             <input_port name="b" required="false"/>
             <input_port name="test" required="false">can be &quot;&lt;&quot;, &quot;&gt;&quot;, or &quot;=&quot;.</input_port>
         </Condition>
+        <Action ID="ComputeFrameAlignment">
+            <input_port name="frameName" required="false"/>
+            <output_port name="out_x" required="false"/>
+            <output_port name="out_y" required="false"/>
+            <output_port name="out_z" required="false"/>
+            <input_port name="x" required="false"/>
+            <input_port name="y" required="false"/>
+            <input_port name="z" required="false"/>
+        </Action>
         <Action ID="ComputeYawAngle">
             <output_port name="output" required="false"/>
             <input_port name="x" required="false"/>
@@ -456,6 +473,10 @@
             <output_port name="dropper2_state" required="false"/>
             <output_port name="torpedo1_state" required="false"/>
             <output_port name="torpedo2_state" required="false"/>
+        </Action>
+        <Action ID="GetBoolTopic">
+            <input_port name="topic" required="true">The topic to subscribe to.</input_port>
+            <output_port name="value" required="false"/>
         </Action>
         <Action ID="GetOdometry">
             <output_port name="op" required="false"/>
